+++ completion_grading/edxapp_wrapper/backends/modulestore_q_v1.py
@@ -1,0 +1,6 @@
+"""
+ModuleStore API module definitions for Open edX Quince release.
+"""
+
+# pylint: disable=unused-import, import-error
+from xmodule.modulestore.django import modulestore